+++ lib/dawn/knowledge_base.rb
@@ -251,6 +251,7 @@ require "dawn/kb/cve_2015_7579"
 
 require "dawn/kb/cve_2016_0751"
 require "dawn/kb/cve_2016_0752"
+require "dawn/kb/cve_2016_0753"
 
 # OSVDB
 
@@ -537,6 +538,7 @@ module Dawn
           Dawn::Kb::CVE_2015_7579.new,
           Dawn::Kb::CVE_2016_0751.new,
           Dawn::Kb::CVE_2016_0752.new,
+          Dawn::Kb::CVE_2016_0753.new,
 
 
           # OSVDB Checks are still here since are all about dependencies